Chore: Move backgroundsvcs and usagestatssvcs to registry
