+++ pkg/registry/usagestatssvcs/usage_stats_providers_registry.go
